requirements.txt
numpy
pandas
matplotlib
beautifulsoup4
ipykernel
openai
googlesearch-python
tqdm